LPD-41335 Merge LanguageExtender into LanguageResourcesExtender, and merge LanguageExtension into LanguageResourcesExtension
